requirements.txt
accelerate==0.17.0.dev0
controlnet-aux==0.0.1
diffusers==0.14.0
huggingface-hub==0.12.1
matplotlib-inline==0.1.6
numpy==1.24.2
opencv-python==4.7.0.72
pickleshare==0.7.5
Pillow==9.4.0
scipy==1.10.1
tokenizers==0.13.2
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
tqdm==4.65.0
transformers==4.26.1
xformers==0.0.16